requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
ollama>=0.1.7
pyyaml>=6.0
pywebview>=5.0
